테슽트 코드 준비중

+++ src/main/java/com/sparta/spring99_homework/dto/RestaurantRequestDto.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 
 public class RestaurantRequestDto {
     private String name;
-    private Long minOrderPrice;
-    private Long deliveryFee;
+    private int minOrderPrice;
+    private int deliveryFee;
 }
